Fix implicit conversion to int
Bug: T351312

+++ includes/ParserFunctions.php
@@ -849,7 +849,7 @@ class ParserFunctions {
 		// Precompute limit to avoid generating enormous string:
 		$diff = mb_strlen( $inReplaceTo ) - mb_strlen( $inReplaceFrom );
 		if ( $diff > 0 ) {
-			$limit = ( ( $this->config->get( 'PFStringLengthLimit' ) - mb_strlen( $inStr ) ) / $diff ) + 1;
+			$limit = (int)( ( $this->config->get( 'PFStringLengthLimit' ) - mb_strlen( $inStr ) ) / $diff ) + 1;
 		} else {
 			$limit = -1;
 		}
